Check tdf's range in set_new_dilation API
1. Since range is [0.001, 100), set_new_dilation takes tdf of type float
number;
2. 1000*TDF is actually echoed to /proc/$pid/dilation file

+++ test_virtual_time/util.h
@@ -25,7 +25,7 @@ int timeval_substract(struct timeval* result, struct timeval* x, struct timeval*
 
 // dilation helpers
 int virtual_time_unshare(int flags); // for historical reason
-int set_new_dilation(pid_t pid, int tdf);
+int set_new_dilation(pid_t pid, float tdf); // NOTICE: 0.001 <= tdf < 100
 int virtual_time_exit(pid_t pid);
 
 // freeze helpers
